feat #105: DailyMenu 분리
